all components updated
