+++ src/Main/Homepage/Homepage.js
@@ -6,15 +6,13 @@ import { Link } from 'react-router-dom';
 
 function Homepage() {
     return (
-     
-     <div className="homepage">
-         <img src={HomepageName} className="homepage-name"/>
-         <p className="press-button-mario">
-             <Link to="/selection">
-             Click Here</Link></p>
-         <img src={HomepageMario} className="homepage-mario"/>
-        
-     </div>
+        <div className="homepage">
+            <img src={HomepageName} className="homepage-name"/>
+            <p className="press-button-mario">
+                <Link to="/selection">Click Here</Link>
+            </p>
+            <img src={HomepageMario} className="homepage-mario"/>
+        </div>
     );
   }
