+++ src/Main/SelectionCharacterPage/SelectionCharacterSection/selectionCharacterSection.css
