--- src/Main/main.css
@@ -1,2 +1,0 @@
-
-
